New 'Find and use data' section from docs team
This PR is to add a ‘how to find data’ section so that customers can find the data they send. I’ve included what I think are the applicable data types (metric and span data) but let me know if there are other data types reported by this exporter. If you think this exporter will change to send other types of data in future, keep in mind it will be important to update this with a link to that data type.